GpgIdentifier: split out into a separate class
